requirements.txt
PyJWT==2.8.0
cryptography==42.0.8
requests==2.32.3
